# routes/analytics.js
const express = require('express');
const router = express.Router();
const { authenticateJWT } = require('../middleware/auth');

// Get productivity analytics
router.get('/productivity', authenticateJWT, async (req, res) => {
  try {
    const { userId, timeframe, period } = req.query;
    
    // TODO: Implement productivity analytics
    const productivityData = {
      userId: userId || req.user._id,
      timeframe: timeframe || 'week',
      period: period || 'current',
      metrics: {
        tasksCompleted: 25,
        averageCompletionTime: '2.3 days',
        efficiency: 87,
        productivityScore: 92
      },
      trends: {
        daily: [85, 88, 90, 87, 92, 89, 91],
        weekly: [87, 89, 91, 88, 90, 92, 89],
        monthly: [88, 89, 90, 91, 92, 89, 90, 91, 92, 89, 90, 91]
      }
    };
    
    // Emit real-time analytics update
    if (req.io) {
      req.io.to(`analytics-${req.user._id}`).emit('productivity-updated', productivityData);
    }
    
    res.json({
      message: 'Productivity analytics retrieved successfully',
      data: productivityData
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch productivity analytics', error: err.message });
  }
});

// Get efficiency analytics
router.get('/efficiency', authenticateJWT, async (req, res) => {
  try {
    const { userId, metric, comparison } = req.query;
    
    // TODO: Implement efficiency analytics
    const efficiencyData = {
      userId: userId || req.user._id,
      metric: metric || 'overall',
      comparison: comparison || 'previous_period',
      currentEfficiency: 89,
      previousEfficiency: 85,
      improvement: 4.7,
      breakdown: {
        timeManagement: 92,
        taskPrioritization: 87,
        focus: 91,
        collaboration: 88
      }
    };
    
    // Emit real-time analytics update
    if (req.io) {
      req.io.to(`analytics-${req.user._id}`).emit('efficiency-updated', efficiencyData);
    }
    
    res.json({
      message: 'Efficiency analytics retrieved successfully',
      data: efficiencyData
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch efficiency analytics', error: err.message });
  }
});

// Get quality analytics
router.get('/quality', authenticateJWT, async (req, res) => {
  try {
    const { userId, qualityMetric, timeframe } = req.query;
    
    // TODO: Implement quality analytics
    const qualityData = {
      userId: userId || req.user._id,
      qualityMetric: qualityMetric || 'overall',
      timeframe: timeframe || 'month',
      qualityScore: 94,
      metrics: {
        accuracy: 96,
        completeness: 92,
        timeliness: 89,
        customerSatisfaction: 95
      },
      trends: {
        weekly: [92, 93, 94, 95, 94, 93, 94],
        monthly: [93, 94, 95, 94, 93, 94, 95, 94, 93, 94, 95, 94]
      }
    };
    
    // Emit real-time analytics update
    if (req.io) {
      req.io.to(`analytics-${req.user._id}`).emit('quality-updated', qualityData);
    }
    
    res.json({
      message: 'Quality analytics retrieved successfully',
      data: qualityData
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch quality analytics', error: err.message });
  }
});

// Get trends analytics
router.get('/trends', authenticateJWT, async (req, res) => {
  try {
    const { userId, trendType, period } = req.query;
    
    // TODO: Implement trends analytics
    const trendsData = {
      userId: userId || req.user._id,
      trendType: trendType || 'productivity',
      period: period || '6_months',
      trends: {
        productivity: {
          direction: 'increasing',
          slope: 0.8,
          confidence: 0.92,
          data: [85, 87, 89, 91, 93, 95]
        },
        efficiency: {
          direction: 'stable',
          slope: 0.1,
          confidence: 0.78,
          data: [88, 89, 88, 89, 88, 89]
        },
        quality: {
          direction: 'increasing',
          slope: 0.6,
          confidence: 0.85,
          data: [90, 91, 92, 93, 94, 95]
        }
      },
      insights: [
        'Productivity shows consistent improvement over the last 6 months',
        'Efficiency remains stable with minor fluctuations',
        'Quality metrics demonstrate steady growth'
      ]
    };
    
    res.json({
      message: 'Trends analytics retrieved successfully',
      data: trendsData
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to fetch trends analytics', error: err.message });
  }
});

// Generate analytics reports
router.post('/reports', authenticateJWT, async (req, res) => {
  try {
    const { reportType, parameters, format, timeframe } = req.body;
    
    // TODO: Implement analytics report generation
    const report = {
      id: Date.now().toString(),
      type: reportType,
      parameters,
      format: format || 'json',
      timeframe: timeframe || 'month',
      generatedAt: new Date().toISOString(),
      downloadUrl: `/api/analytics/reports/download/${Date.now()}`,
      summary: {
        totalMetrics: 15,
        keyInsights: 3,
        recommendations: 5
      }
    };
    
    res.json({
      message: 'Analytics report generated successfully',
      report
    });
  } catch (err) {
    res.status(500).json({ message: 'Failed to generate analytics report', error: err.message });
  }
});

module.exports = router; 