# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "AI Task Manager API",
    "version": "1.0.0",
    "description": "Comprehensive API documentation for the AI Task Manager SaaS platform with task management, AI features, team management, gamification, analytics, and more."
  },
  "servers": [
    { "url": "http://localhost:3000/api" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "ProductivityReport": {
        "type": "object",
        "properties": {
          "userId": { "type": "string", "description": "User ID for the report" },
          "timeframe": { "type": "string", "enum": ["week", "month", "quarter", "year"], "description": "Report timeframe" },
          "metrics": { "type": "string", "enum": ["all", "efficiency", "quality", "collaboration"], "description": "Metrics included" },
          "productivityScore": { "type": "number", "description": "Overall productivity score" },
          "efficiency": { "type": "number", "description": "Efficiency rating" },
          "quality": { "type": "number", "description": "Quality rating" },
          "trends": {
            "type": "object",
            "properties": {
              "weekly": { "type": "array", "items": { "type": "number" }, "description": "Weekly productivity trends" },
              "monthly": { "type": "array", "items": { "type": "number" }, "description": "Monthly productivity trends" }
            }
          },
          "recommendations": { "type": "array", "items": { "type": "string" }, "description": "Productivity improvement recommendations" }
        }
      },
...
    }
  },
  "security": [
    { "bearerAuth": [] }
  ],
...
  "paths": {
...
    "/analytics/productivity": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get productivity analytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string", "description": "User ID to filter analytics" } },
          { "name": "timeframe", "in": "query", "schema": { "type": "string", "enum": ["day", "week", "month", "quarter", "year"], "description": "Timeframe for analytics" } },
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["current", "previous", "custom"], "description": "Period for analytics" } }
        ],
        "responses": {
          "200": { "description": "Productivity analytics retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/analytics/efficiency": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get efficiency analytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string", "description": "User ID to filter analytics" } },
          { "name": "metric", "in": "query", "schema": { "type": "string", "enum": ["overall", "time_management", "task_prioritization", "focus", "collaboration"], "description": "Efficiency metric to analyze" } },
          { "name": "comparison", "in": "query", "schema": { "type": "string", "enum": ["previous_period", "baseline", "target"], "description": "Comparison type" } }
        ],
        "responses": {
          "200": { "description": "Efficiency analytics retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/analytics/quality": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get quality analytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string", "description": "User ID to filter analytics" } },
          { "name": "qualityMetric", "in": "query", "schema": { "type": "string", "enum": ["overall", "accuracy", "completeness", "timeliness", "customer_satisfaction"], "description": "Quality metric to analyze" } },
          { "name": "timeframe", "in": "query", "schema": { "type": "string", "enum": ["week", "month", "quarter", "year"], "description": "Timeframe for analytics" } }
        ],
        "responses": {
          "200": { "description": "Quality analytics retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/analytics/trends": {
      "get": {
        "tags": ["Analytics"],
        "summary": "Get trends analytics",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string", "description": "User ID to filter analytics" } },
          { "name": "trendType", "in": "query", "schema": { "type": "string", "enum": ["productivity", "efficiency", "quality", "overall"], "description": "Type of trend to analyze" } },
          { "name": "period", "in": "query", "schema": { "type": "string", "enum": ["3_months", "6_months", "1_year", "custom"], "description": "Period for trend analysis" } }
        ],
        "responses": {
          "200": { "description": "Trends analytics retrieved successfully" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
    "/analytics/reports": {
      "post": {
        "tags": ["Analytics"],
        "summary": "Generate analytics reports",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reportType": { "type": "string", "enum": ["productivity", "efficiency", "quality", "comprehensive"], "description": "Type of report to generate" },
                  "parameters": { "type": "object", "description": "Report parameters" },
                  "format": { "type": "string", "enum": ["json", "pdf", "csv", "excel"], "description": "Report format" },
                  "timeframe": { "type": "string", "enum": ["week", "month", "quarter", "year"], "description": "Report timeframe" }
                },
                "required": ["reportType"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Analytics report generated successfully" },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
    "/api/reports/productivity": {
      "get": {
        "tags": ["Reports"],
        "summary": "Get productivity reports",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "userId", "in": "query", "schema": { "type": "string" }, "description": "User ID to filter productivity data" },
          { "name": "timeframe", "in": "query", "schema": { "type": "string", "enum": ["week", "month", "quarter", "year"], "default": "month" }, "description": "Report timeframe" },
          { "name": "metrics", "in": "query", "schema": { "type": "string", "enum": ["all", "efficiency", "quality", "collaboration"], "default": "all" }, "description": "Specific metrics to include" }
        ],
        "responses": {
          "200": {
            "description": "Productivity report generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "report": { "$ref": "#/components/schemas/ProductivityReport" }
                  }
                }
              }
            }
          },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Internal server error" }
        }
      }
    },
...
  }
}